♻️ Refactor: Home/ProfilePage 코드 스타일 수정 및 svg 변경
