+++ src/pages/UserProfilePage.jsx
@@ -1,8 +1,9 @@
 import styled from 'styled-components';
-import UserProfileHeader from '../components/Profile/UserProfileHeader';
-import ProductsContainer from '../components/Profile/ProductsContainer';
-import PostContainer from '../components/Profile/PostContainer';
 import Header from '../components/Header/Header';
+import ProfileInfo from '../components/Profile/ProfileInfo';
+import ProfileProducts from '../components/Profile/ProfileProducts';
+import ProfilePosts from '../components/Profile/ProfilePosts';
+import { useParams } from 'react-router-dom';
 
 const Container = styled.main`
   display: flex;
@@ -14,15 +15,16 @@ const Container = styled.main`
   gap: 6px;
 `;
 
-const UserProfilePage = () => {
+export default function MyProfilePage() {
+  const { _id } = useParams();
+  console.log(_id);
   return (
     <Container>
       <Header type="basic" />
       <h1 className="a11y-hidden">나의 프로필 페이지</h1>
-      <UserProfileHeader />
-      <ProductsContainer />
-      <PostContainer />
+      <ProfileInfo user={_id} />
+      <ProfileProducts />
+      <ProfilePosts />
     </Container>
   );
-};
-export default UserProfilePage;
+}
